pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "bfportal_grpc"
version = "0.1.4"
description = " bf2042 portal web grpc as python package "
readme = "Readme.md"
authors = ["iiTzArcur <arcur@gametools.network>"]
# license = { file = "LICENSE" }
classifiers = [
    # "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["bfportal", "grpc", "Battlefield portal"]
repository = "https://github.com/community-network/bfportal-grpc"
packages = [{ include = "bfportal_grpc" }]

[tool.poetry.dependencies]
aiohttp = "^3.8.5"
protobuf = "^4.24.3"
grpcio = "^1.58"
python = "^3.10"
proto-compile = "^0.1.7"
pytest-asyncio = "^0.21.1"
python-dotenv = "^1.0.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/community-network/bfportal-grpc/issues"

[tool.poetry.scripts]
compile-proto = "bfportal_grpc.helpers.compile:compile_proto"


[tool.poetry.group.dev.dependencies]
pytest = "*"
sonora = "*"
black = "*"
isort = "*"
grpcio-tools = "^1.59.0"

[tool.poetry.build]
generate-setup-file = false
script = "bfportal_grpc/helpers/compile.py"

[tool.isort]
profile = "black"
import_heading_stdlib = "Standard library imports"
import_heading_thirdparty = "Third party imports"
import_heading_firstparty = "Reader imports"

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "feedparser"
ignore_missing_imports = true
